+++ goodrouter-ts/src/main.ts
@@ -1,3 +1,3 @@
-export * from "./route-node-json.js";
+export * from "./json.js";
 export * from "./router-options.js";
 export * from "./router.js";
