--- goodrouter-ts/src/route-node-json.ts
@@ -1,6 +1,0 @@
-export interface RouteNodeJson<K extends string | number> {
-    anchor: string;
-    hasParameter: boolean;
-    routeKey: K | null;
-    children: RouteNodeJson<K>[];
-}
