+++ goodrouter-ts/src/router.spec.ts
@@ -1,5 +1,6 @@
 import assert from "node:assert/strict";
 import test from "node:test";
+import { RouterJson } from "./json.js";
 import { Router } from "./router.js";
 import { parametersFromTemplates } from "./testing/parameters.js";
 import { loadTemplates } from "./testing/templates.js";
@@ -155,6 +156,46 @@ test("router bug", () => {
         ),
         ["c", { enterprise: "xx", runner_group_id: "yy" }],
     );
+});
+
+test("route-node-json", () => {
+    const router = new Router();
+    router.insertRoute(1, "x/y");
+    router.insertRoute(2, "x/z");
+    const actual = router.saveToJson();
+    const expected: RouterJson<number> = {
+        templatePairs: [
+            [1, [["x/y", null]]],
+            [2, [["x/z", null]]],
+        ],
+        rootNode: {
+            anchor: "",
+            hasParameter: false,
+            routeKey: null,
+            children: [
+                {
+                    anchor: "x/",
+                    hasParameter: false,
+                    routeKey: null,
+                    children: [
+                        {
+                            anchor: "y",
+                            hasParameter: false,
+                            routeKey: 1,
+                            children: [],
+                        },
+                        {
+                            anchor: "z",
+                            hasParameter: false,
+                            routeKey: 2,
+                            children: [],
+                        },
+                    ],
+                },
+            ],
+        },
+    };
+    assert.deepEqual(actual, expected);
 });
 
 testTemplates("small");
